% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featselect.R
\name{featselectRF}
\alias{featselectRF}
\title{Feature selection algorithm for Random Forests (RF)}
\usage{
featselectRF(x, y, nRepeat = 100, kFold = 5, rKeep = 0.3,
  bParallel = TRUE, nThread = parallel::detectCores() - 1, nSeed = 1983,
  pCutOff = 0.05, ...)
}
\arguments{
\item{x}{N x D predictors data frame where N - no. of samples, D - no. of features}

\item{y}{a vector of factors of length N, the target class (e.g as.factor("A", "A", "B", etc.))}

\item{nRepeat}{no. of times features are permuted (this is the sample size used when comparing importance
score for permuted vs real features)}

\item{kFold}{no. of cross-validation folds (default: 5)}

\item{rKeep}{percentage of predictors to ignore after first RF fit (0>rKeep<=1) (default: 0.3)}

\item{bParallel}{whether to use mProbes() or mProbesParallel() (default: True)}

\item{nThread}{no. of threads to use when running in parallel (default: parallel::detected cores - 1)}

\item{nSeed}{seed for cross-validation folds (default: 1983)}

\item{pCutOff}{Bonferonni corrected adjusted p-value cutoff when comparing importance scores of 
permuted vs real predictors (default: 0.05)}

\item{...}{arguments passed to the Random Forest classifier (e.g ntree, sampsize, etc.)}
}
\value{
A list with the following components:
\item{rKeepPredictors}{rKeep\% predictors kept after first RF fit}
\item{topPredictors }{top predictors (Bonferroni corrected p-values<pCutOff) of each fold}
\item{pValues}{Bonferroni corrected pValues for rKeepPredictors}
\item{ROC}{receiver operating characteristic curve, ROCR object}
\item{auc}{area under the ROC curve}
\item{confMatrix}{confusion matrix on test data}
\item{iiFolds}{indices of the cross-validation folds}
}
\description{
Implements a modified mProbes/xRF feature selection algorithm within a
cross-validation loop
}
\examples{
bWant <- iris$Species \%in\% c("versicolor", "virginica")
x <- iris[bWant, 1:4]
y <- droplevels(as.factor(iris$Species[bWant]))

out <- featselectRF(x, y, nodesize=3, ntree=1001)

}
\references{
\href{https://doi.org/10.1093/bioinformatics/bts238}{Huynh-Thu VA et al. Bioinformatics 2012}

\href{http://dx.doi.org/10.1155/2015/471371}{Nguyen et al. The Scientific World Journal 2015}
}
