% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featselect.R
\name{mProbes}
\alias{mProbes}
\title{mProbes feature selection algorithm}
\usage{
mProbes(x, y, nRepeat = 100, ...)
}
\arguments{
\item{x}{N x D predictors data frame where N - no. of samples, D - no. of features}

\item{y}{a vector of factors of length N, the target class (e.g as.factor("A", "A", "B", etc.))}

\item{nRepeat}{no. of times features are permuted (this is the sample size used when comparing importance
score for permuted vs real features)}

\item{...}{arguments passed to the Random Forest classifier (e.g ntree, sampsize, etc.)}
}
\value{
A list with the following components:
\item{impMetric}{2D x nRepeat matrix of variable importance measures for each predictor (permuted and not) 
for every repeat (Note: the permuted variables have the suffix "Perm")}
\item{FWER}{a numeric vector of length D with the family wise error rate computed for every feature}
}
\description{
Implements the mProbes feature selection algorithm for Random Forests
}
\examples{
bWant <- iris$Species \%in\% c("versicolor", "virginica")
x <- iris[bWant, 1:4]
y <- droplevels(as.factor(iris$Species[bWant]))

out <- mProbes(x, y, 100)

}
\references{
\href{https://academic.oup.com/bioinformatics/article/28/13/1766/234473/Statistical-interpretation-of-machine-learning}{Huynh-Thu VA et al. Bioinformatics 2012}
}
