% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featselect.R
\name{mProbesParallel}
\alias{mProbesParallel}
\title{mProbes feature selection algorithm parallelised}
\usage{
mProbesParallel(x, y, nRepeat = 100, nThread = parallel::detectCores() - 1,
  ...)
}
\arguments{
\item{x}{N x D predictors data frame where N - no. of samples, D - no. of features}

\item{y}{a vector of factors of length N, the target class (e.g as.factor("A", "A", "B", etc.))}

\item{nRepeat}{no. of times features are permuted (this is the sample size used when comparing importance
score for permuted vs real features)}

\item{nThread}{no. of threads to use to run it in parallel (default: parallel::detected cores - 1)}

\item{...}{arguments passed to the Random Forest classifier (e.g ntree, sampsize, etc.)}
}
\value{
A list with the following components:
\item{impMetric}{2D x nRepeat matrix of variable importance measures for each predictor (permuted and not) 
for every repeat (Note: the permuted variables have the suffix "Perm")}
\item{FWER}{a numeric vector of length D with the family wise error rate computed for every feature}
}
\description{
Same as \code{\link{mProbes}} but parallelised
}
\examples{
bWant <- iris$Species \%in\% c("versicolor", "virginica")
x <- iris[bWant, 1:4]
y <- droplevels(as.factor(iris$Species[bWant]))

out <- mProbesParallel(x, y, 100, 4)

}
\references{
\href{https://doi.org/10.1093/bioinformatics/bts238}{Huynh-Thu VA et al. Bioinformatics 2012}
}
