% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotconfusionmatrix}
\alias{plotconfusionmatrix}
\title{Plot 2 x 2 confusion matrix}
\usage{
plotconfusionmatrix(confMatrix, plotLabels = colnames(confMatrix)[1:2],
  plotTitle = "", colDiag = "darkblue", colOffDiag = "grey",
  colAxes = "blue", colTicks = "red")
}
\arguments{
\item{confMatrix}{a 2 x 2 confusion matrix (classA | classB | classError)
predicted vs observed}

\item{plotLabels}{a character vector of length two with the ordered name of the
two classes (default: colnames(confMatrix)[1:2])}

\item{plotTitle}{main title of plot (default: "")}

\item{colDiag}{colour of the diagonal segments (correctly classified) 
(default: "darkblue")}

\item{colOffDiag}{colour of the off diagnonal segments (misclassified)
(default: "grey")}

\item{colAxes}{colour of the x-y axes (default: "blue")}

\item{colTicks}{colour of the x-y ticks (default: "red")}
}
\value{
None, a plot is created
}
\description{
This function creates an illustrated confusion matrix for two-class problems
}
