% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotROC}
\alias{plotROC}
\title{Plot ROC curves}
\usage{
plotROC(ROC, AUC, colAvg = "black", colFold = "grey", colRandom = "red")
}
\arguments{
\item{ROC}{a list of ROC curves produced by ROCR and named ("Fold1, Fold2,...,Average")}

\item{AUC}{a list of area under the ROC curve (same size as ROC)}

\item{colAvg}{colour for the average ROC curve}

\item{colFold}{colour for each fold ROC curve}

\item{colRandom}{colour for randomly guession (y=x) line}
}
\value{
None, a plot is created
}
\description{
This function plots ROC curves for each fold and their average
}
